Revert "feat: add late-load mode support (https://github.com/tiann/KernelSU/pull/3258)"
This reverts commit 948f47715d46232735ebb174c26544c87eeff64a.
