+++ kernel/ksu.c
@@ -2,20 +2,17 @@
 #include <linux/fs.h>
 #include <linux/kobject.h>
 #include <linux/module.h>
-#include <linux/sched.h>
 #include <linux/workqueue.h>
 
 #include "allowlist.h"
 #include "feature.h"
 #include "klog.h" // IWYU pragma: keep
-#include "manager.h"
 #include "throne_tracker.h"
 #include "syscall_hook_manager.h"
 #include "ksud.h"
 #include "supercalls.h"
 #include "ksu.h"
 #include "file_wrapper.h"
-#include "selinux/selinux.h"
 
 // workaround for A12-5.10 kernel
 // Some third-party kernel (e.g. linegaeOS) uses wrong toolchain, which supports
@@ -55,16 +52,9 @@ __attribute__((naked)) int __init kernelsu_init_early(void)
 #endif
 
 struct cred *ksu_cred;
-bool ksu_late_loaded;
 
 int __init kernelsu_init(void)
 {
-#ifdef MODULE
-	ksu_late_loaded = (current->pid != 1);
-#else
-	ksu_late_loaded = false;
-#endif
-
 #ifdef CONFIG_KSU_DEBUG
 	pr_alert("*************************************************************");
 	pr_alert("**     NOTICE NOTICE NOTICE NOTICE NOTICE NOTICE NOTICE    **");
@@ -84,36 +74,15 @@ int __init kernelsu_init(void)
 
 	ksu_supercalls_init();
 
-	if (ksu_late_loaded) {
-		pr_info("late load mode, skipping kprobe hooks\n");
+	ksu_syscall_hook_manager_init();
 
-		apply_kernelsu_rules();
-		cache_sid();
-		setup_ksu_cred();
+	ksu_allowlist_init();
 
-		ksu_allowlist_init();
-		ksu_load_allow_list();
+	ksu_throne_tracker_init();
 
-		ksu_syscall_hook_manager_init();
+	ksu_ksud_init();
 
-		ksu_throne_tracker_init();
-		ksu_observer_init();
-		ksu_file_wrapper_init();
-
-		ksu_boot_completed = true;
-		track_throne(false);
-
-	} else {
-		ksu_syscall_hook_manager_init();
-
-		ksu_allowlist_init();
-
-		ksu_throne_tracker_init();
-
-		ksu_ksud_init();
-
-		ksu_file_wrapper_init();
-	}
+    ksu_file_wrapper_init();
 
 #ifdef MODULE
 #ifndef CONFIG_KSU_DEBUG
@@ -132,8 +101,7 @@ void kernelsu_exit(void)
 
 	ksu_observer_exit();
 
-	if (!ksu_late_loaded)
-		ksu_ksud_exit();
+	ksu_ksud_exit();
 
 	ksu_syscall_hook_manager_exit();
