+++ kernel/ksu.h
@@ -27,6 +27,5 @@ static inline int endswith(const char *s, const char *t)
 }
 
 extern struct cred *ksu_cred;
-extern bool ksu_late_loaded;
 
 #endif
