+++ manager/app/src/main/cpp/ksu.cc
@@ -101,23 +101,15 @@ bool is_safe_mode() {
 bool is_lkm_mode() {
     auto info = get_info();
     if (info.version > 0) {
-        return (info.flags & KSU_GET_INFO_FLAG_LKM) != 0;
+        return (info.flags & 0x1) != 0;
     }
-    return (legacy_get_info().second & KSU_GET_INFO_FLAG_LKM) != 0;
-}
-
-bool is_late_load_mode() {
-    auto info = get_info();
-    if (info.version > 0) {
-        return (info.flags & KSU_GET_INFO_FLAG_LATE_LOAD) != 0;
-    }
-    return false;
+    return (legacy_get_info().second & 0x1) != 0;
 }
 
 bool is_manager() {
     auto info = get_info();
     if (info.version > 0) {
-        return (info.flags & KSU_GET_INFO_FLAG_MANAGER) != 0;
+        return (info.flags & 0x2) != 0;
     }
     return legacy_get_info().first > 0;
 }
