+++ manager/app/src/main/java/com/twj/wksu/Natives.kt
@@ -45,9 +45,6 @@ object Natives {
     val isLkmMode: Boolean
         external get
 
-    val isLateLoadMode: Boolean
-        external get
-
     val isManager: Boolean
         external get
