+++ manager/app/src/main/res/values/strings.xml
@@ -76,7 +76,6 @@
     <string name="hide_system_apps">Hide system apps</string>
     <string name="export_log">Export logs</string>
     <string name="safe_mode">Safe mode</string>
-    <string name="jailbreak_mode">Jailbreak mode</string>
     <string name="reboot_to_apply">Reboot to take effect</string>
     <string name="module_magisk_conflict">Modules are unavailable due to a conflict with Magisk!</string>
     <string name="home_mount_system">Mount system</string>
