+++ userspace/ksud/src/init_event.rs
@@ -117,7 +117,7 @@ pub fn on_post_data_fs() -> Result<()> {
     Ok(())
 }
 
-pub fn run_stage(stage: &str, block: bool) {
+fn run_stage(stage: &str, block: bool) {
     utils::umask(0);
 
     if utils::has_magisk() {
