--- userspace/ksud/src/late_load.rs
@@ -1,88 +1,0 @@
-use anyhow::{Context, Result};
-use log::{info, warn};
-
-use crate::module::{handle_updated_modules, prune_modules};
-use crate::{assets, defs, init_event, metamodule, restorecon, utils};
-
-pub fn run() -> Result<()> {
-    info!("late-load command triggered!");
-
-    // 1. Check if KernelSU is already loaded
-    if ksuinit::has_kernelsu() {
-        info!("KernelSU already loaded, skip loading ko");
-    } else {
-        // 2. Detect current KMI version
-        let kmi =
-            crate::boot_patch::get_current_kmi().context("Failed to detect current KMI version")?;
-        info!("Detected KMI: {kmi}");
-
-        // 3. Get kernelsu.ko from embedded assets
-        let ko_name = format!("{kmi}_kernelsu.ko");
-        let ko_data = assets::get_asset_data(&ko_name)
-            .with_context(|| format!("Failed to get {ko_name} from assets"))?;
-
-        // 4. Load kernelsu.ko from memory with manual relocation
-        info!("Loading kernelsu.ko for KMI {kmi}...");
-        ksuinit::load_module(&ko_data).context("Failed to load kernelsu.ko")?;
-        info!("kernelsu.ko loaded successfully!");
-    }
-
-    utils::umask(0);
-
-    if let Err(e) = crate::module_config::clear_all_temp_configs() {
-        warn!("clear temp configs failed: {e}");
-    }
-
-    utils::install(None).context("Failed to install ksud")?;
-
-    // 5. Handle module updates
-    if let Err(e) = handle_updated_modules() {
-        warn!("handle updated modules failed: {e}");
-    }
-
-    if let Err(e) = prune_modules() {
-        warn!("prune modules failed: {e}");
-    }
-
-    if let Err(e) = restorecon::restorecon() {
-        warn!("restorecon failed: {e}");
-    }
-
-    // 6. Load SELinux rules
-    if crate::module::load_sepolicy_rule().is_err() {
-        warn!("load sepolicy.rule failed");
-    }
-
-    if let Err(e) = crate::profile::apply_sepolies() {
-        warn!("apply root profile sepolicy failed: {e}");
-    }
-
-    // 7. Initialize features
-    if let Err(e) = crate::feature::init_features() {
-        warn!("init features failed: {e}");
-    }
-
-    // 8. Execute late-load stage scripts (blocking)
-    init_event::run_stage("late-load", true);
-
-    // 9. Load system.prop
-    if let Err(e) = crate::module::load_system_prop() {
-        warn!("load system.prop failed: {e}");
-    }
-
-    // 10. Execute metamodule mount script (OverlayFS)
-    if let Err(e) = metamodule::exec_mount_script(defs::MODULE_DIR) {
-        warn!("execute metamodule mount failed: {e}");
-    }
-
-    // 11. Execute post-mount stage scripts (blocking)
-    init_event::run_stage("post-mount", true);
-
-    // 12. Execute service stage scripts (non-blocking)
-    init_event::run_stage("service", false);
-
-    // 13. Execute boot-completed stage scripts (non-blocking)
-    init_event::run_stage("boot-completed", false);
-
-    Ok(())
-}
