+++ userspace/ksud/src/main.rs
@@ -27,8 +27,6 @@ mod init_event;
 #[cfg(target_os = "android")]
 mod ksucalls;
 #[cfg(target_os = "android")]
-mod late_load;
-#[cfg(target_os = "android")]
 mod metamodule;
 #[cfg(target_os = "android")]
 mod module;
