+++ userspace/ksuinit/Cargo.toml
@@ -12,6 +12,8 @@ goblin = "0.10"
 scroll = "0.13"
 
 anyhow = "1"
+# If you want to use the following dependencies, please use aarch64-unknown-linux-musl & x86_64-unknown-linux-musl to compile statically
+# rustix = { git = "https://github.com/bytecodealliance/rustix", rev = "7b44528", features = ["mount", "fs", "runtime", "system", "process"] }
 rustix = { git = "https://github.com/Kernel-SU/rustix.git", rev = "4a53fbc", features = ["mount", "fs", "runtime", "system", "process"] }
 
 syscalls = { version = "0.7", default-features = false, features = [
