+++ userspace/ksuinit/src/main.rs
@@ -1,6 +1,7 @@
 #![no_main]
 
 mod init;
+mod loader;
 
 use rustix::{cstr, runtime::execve};
 /// # Safety
