+++ website/docs/ru_RU/guide/metamodule.md
@@ -422,74 +422,6 @@ fsconfig_set_string(fs, "source", "KSU")?;  // ОБЯЗАТЕЛЬНО!
 
 Нет. Он обеспечивает стандартное монтирование overlayfs, совместимое с большинством модулей. Вы можете создать свой собственный метамодуль, если вам нужно другое поведение.
 
-## Режим late-load {#late-load-mode}
-
-Помимо стандартного процесса загрузки, описанного выше, KernelSU поддерживает **режим late-load** для сценариев LKM (загружаемый модуль ядра). В этом режиме модуль ядра KernelSU загружается **после полной загрузки системы**, а не во время процесса init.
-
-### Когда происходит late-load?
-
-Late-load запускается командой `ksud late-load`. Эта команда:
-
-1. Определяет текущую версию KMI и загружает соответствующий `kernelsu.ko` из встроенных ресурсов.
-2. Выполняет инициализацию модуля (правила SELinux, список разрешений, функции и т.д.), которая обычно происходит во время загрузки.
-
-Поскольку система уже полностью запущена, некоторые механизмы времени загрузки недоступны или не нужны.
-
-### Отличия от стандартной загрузки
-
-| Поведение | Стандартная загрузка | Режим late-load |
-|-----------|:---:|:---:|
-| Модуль ядра загружен init (PID 1) | Да | Нет (загружен после загрузки) |
-| Хуки kprobe ksud (execve/read/fstat/input) | Да | Пропущены |
-| Обнаружение безопасного режима (клавиша громкости) | Да | Всегда отключено |
-| Захват журнала загрузки (logcat/dmesg) | Да | Пропущен |
-| Проверка сосуществования с Magisk | Да | Пропущена |
-| Событие `post-fs-data` отправлено ядру | Да | Пропущено |
-| Событие `boot-completed` отправлено ядру | Да | Установлено напрямую при инициализации |
-| Скрипты `post-fs-data.sh` / `post-fs-data.d/` | Да | Заменены этапом `late-load` |
-| Загрузка `system.prop` | Да | Да |
-| Монтирование OverlayFS (metamodule) | Да | Да |
-| Скрипты `post-mount.sh` / `post-mount.d/` | Да | Да |
-| Скрипты `service.sh` / `service.d/` | Да | Да |
-| Скрипты `boot-completed.sh` / `boot-completed.d/` | Да | Да |
-| Переменная окружения `KSU_LATE_LOAD` | Не установлена | Установлена в `1` |
-| Флаг info ядра `0x4` | Не установлен | Установлен |
-
-### Порядок выполнения скриптов
-
-В режиме late-load порядок выполнения скриптов следующий:
-
-```txt
-ksud late-load:
-  1. Загрузить kernelsu.ko (если ещё не загружен)
-  2. Извлечь бинарные файлы, обработать обновления модулей, загрузить правила SELinux, инициализировать функции
-  3. Выполнить скрипты late-load.d/ и скрипты late-load модулей (блокирующе)
-  4. Загрузить system.prop (resetprop -n)
-  5. Выполнить скрипт монтирования metamodule (OverlayFS)
-  6. Выполнить скрипты post-mount.d/ и post-mount.sh модулей (блокирующе)
-  7. Выполнить скрипты service.d/ и service.sh модулей (неблокирующе)
-  8. Выполнить скрипты boot-completed.d/ и boot-completed.sh модулей (неблокирующе)
-```
-
-### Скрипты, специфичные для late-load
-
-Модули могут предоставить скрипт `late-load.sh`, который выполняется **только** в режиме late-load, как замена `post-fs-data.sh`. Этот скрипт выполняется до монтирования OverlayFS, аналогично `post-fs-data.sh` в стандартном потоке.
-
-Кроме того, общие скрипты можно размещать в `/data/adb/late-load.d/` для выполнения на этом этапе.
-
-### Обнаружение режима late-load в скриптах
-
-Модули могут определить режим late-load, проверив переменную окружения `KSU_LATE_LOAD`:
-
-```sh
-if [ "$KSU_LATE_LOAD" = "1" ]; then
-    # Работа в режиме late-load
-    echo "Late-load mode detected"
-fi
-```
-
-Это позволяет модулям соответствующим образом корректировать своё поведение, например, пропуская операции, необходимые только при ранней загрузке.
-
 ## См. Также
 
 - [Руководство по Модулям](module.md) - Общая разработка модулей
